requirements.txt
langchain
langchain_community
langchain-google-genai
faiss-cpu
streamlit
sentence_transformers
dotenv-python